Synchronize Port Bind Operation
Currently the proton-server does not wait for the bind operation in the shim
layer to complete. Also, the vif information from the bind operation is not
updated in the baseport object.

This commit contains changes to use etcd to synchronize the bind operation
between the proton-server and the shim layer server.  An example shim
layer server is provided to show how the interaction works.  A few minor
bugs were fixed as well.

Change-Id: Iec1feb2ae4a43afb9e1fb70881603a4a1d1faf68
Implements: blueprint proton-shim-sync