First draft of port and service binding model for Gluon
This is the first draft of a generic port and service-binding model
for Gluon and Proton. Its goal is to define how new networking
service APIs can be modeled and bound to Proton baseports.

Moreover, this document shall support developers of new APIs
by providing a guideline for new APIs.

In addition to the generic model, a slightly adapted version of the
L3VPN model is included as an example.

Implements: blueprint port-service-binding-model

Change-Id: If59e0acd8e36552f761e52256e5af278efea5773
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>