Port and Service Binding Model for Gluon
This document describes the generic port and service-binding model
available in Gluon to define new networking service APIs.

Moreover, this document shall support developers of new APIs
by providing a guideline for new APIs.  In addition to the generic
model, a slightly adapted version of the L3VPN model is included as
an example.  This commit is a combination of:

https://review.openstack.org/392250 and
https://review.openstack.org/395738

Co-Authored-By: Georg.Kunz@ericsson.com
Co-Authored-By: Thomas.Hambleton@nokia.com

Change-Id: I69f2b530088d4dd7bc21be502210db1740f507d3
Implements: blueprint port-service-binding-model