Fix unhandled etcd read exception
The core plugin wrapper code did not handle an exception during an etcd read
of the /gluon directory. This can happen on a new system when there are no
objects created yet. This resulted in a stack trace showing up in the neutron
log file.

The code was modified to handle the execption gracefully.

Change-Id: I8dd34a86454d145af03a0d71590de5ad4fa54fcb
Closes-Bug: #1651599