Using config option to bind to etcd server
In OPNFV Apex deployments, the etcd server listens only at a
specific host IP address, not localhost. As a result, the Gluon
plugin cannot talk to etcd in such deployments.

This fix (mis-)uses the Neutron bind_host configuration option
to allow the Gluon ML2 plugin wrapper to idenify the IP etcd
is listening on. This should be replaced by a proper config
option later on.

Change-Id: Ib9d958c267d0364889e1213f9fc7c93b4fd04d89
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>