Adding initial version of the SFC to Gluon
This is an initial version of a Gluon API model for the IETF SFC
model.

Note: This is is a simplified version targeting the Gluon demo at the
Boston OpenStack summit. It does not fully represent the IETF SFC model.
The model is supposed to be extended in future releases.

Change-Id: I243f10be4f39baa6bce767283070e2a4915ef93a
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>