[WIP] Add ipaddress attribute to Port Object
The current design requires to look up the binding table for the
ipaddress of a port. By adding the ipaddress attribute to the port
object, we only need to look up the port table get the ipaddress and
other port attributes.
Initially this attribute has no value when a port is first created.
A value will be assign to it during binding.

Change-Id: Idd4803cd67f16b3221f934a4c9dbb1a4765a8b5e