Add ipaddress to BaseServiceBinding
If we need to add ip address when we are doing bindings,
we should add ip address as an attribute to BaseServiceBinding.
By adding it to the BaseServiceBinding, we will enforce all bindings
to have a consistent name across all services.

For example, in ietf-sfc we have both sfDataPlaneLocator and
sffDataPlaneLocator which extends BaseServiceBinding.They named their
ip address attribute “ip”. For net-l3vpn, they have vpnBinding extends
BaseServiceBinding, and it name its ip address attribute “ipaddress”.
This inconsistence in naming the ip attribute causes some troubles
in auto generating the backend.

Change-Id: I9f3f53937ae25e5e1dbfc5939db18d476c137f86